Adds sanity check for uploaded workpad JSON (#24297)
* Extracted upload handler. Added sanity check for uploaded JSON

* Changed file name and error verbiage

* Fixed import

* Added early return for null file

* Cleaned up update_workpad
