+++ x-pack/plugins/canvas/public/components/workpad_loader/workpad_dropzone/index.js
@@ -6,30 +6,13 @@
 
 import PropTypes from 'prop-types';
 import { compose, withHandlers } from 'recompose';
-import { getId } from '../../../lib/get_id';
 import { notify } from '../../../lib/notify';
+import { uploadWorkpad } from '../upload_workpad';
 import { WorkpadDropzone as Component } from './workpad_dropzone';
 
 export const WorkpadDropzone = compose(
   withHandlers({
-    onDropAccepted: ({ onUpload }) => ([file]) => {
-      // TODO: Clean up this file, this loading stuff can, and should be, abstracted
-      const reader = new FileReader();
-
-      // handle reading the uploaded file
-      reader.onload = () => {
-        try {
-          const workpad = JSON.parse(reader.result);
-          workpad.id = getId('workpad');
-          onUpload(workpad);
-        } catch (e) {
-          notify.error(e, { title: `Couldn't upload '${file.name || 'file'}'` });
-        }
-      };
-
-      // read the uploaded file
-      reader.readAsText(file);
-    },
+    onDropAccepted: ({ onUpload }) => ([file]) => uploadWorkpad(file, onUpload),
     onDropRejected: () => ([file]) => {
       notify.warning('Only JSON files are accepted', {
         title: `Couldn't upload '${file.name || 'file'}'`,
